Move internal model stuff out of elasticsearch
