+++ common/internal_model/docker-compose.yml
@@ -1,0 +1,11 @@
+elasticsearch:
+  image: "docker.elastic.co/elasticsearch/elasticsearch:7.9.0"
+  ports:
+    - "9200:9200"
+    - "9300:9300"
+  environment:
+    - "http.host=0.0.0.0"
+    - "transport.host=0.0.0.0"
+    - "cluster.name=wellcome"
+    - "ES_JAVA_OPTS=-Xms750m -Xmx750m"
+    - "discovery.type=single-node"
