+++ common/internal_model/src/main/scala/uk/ac/wellcome/models/index/ImagesIndexConfig.scala
@@ -1,9 +1,10 @@
-package uk.ac.wellcome.elasticsearch
+package uk.ac.wellcome.models.index
 
 import com.sksamuel.elastic4s.ElasticDsl.{keywordField, _}
 import com.sksamuel.elastic4s.analysis.Analysis
 import com.sksamuel.elastic4s.requests.mappings.{MappingDefinition, ObjectField}
 import com.sksamuel.elastic4s.requests.mappings.dynamictemplate.DynamicMapping
+import uk.ac.wellcome.elasticsearch.{IndexConfig, IndexConfigFields, WorksAnalysis}
 
 object InitialImageIndexConfig extends IndexConfig {
